better subprocess handling and other check of homebrew cask being installed as caskroom moved to homebrew

+++ brew-cask-upgrade.sh
@@ -732,21 +732,55 @@ echo ''
 echo -e "\033[1mupdating homebrew, formulas and casks...\033[0m"
 echo ''
 
+function get_running_subprocesses()
+{
+    SUBPROCESSES_PID_TEXT=$(pgrep -lg $(ps -o pgid= $$) | grep -v $$ | grep -v grep)
+    SCRIPT_COMMAND=$(ps -o comm= $$)
+	PARENT_SCRIPT_COMMAND=$(ps -o comm= $PPID)
+	if [[ $PARENT_SCRIPT_COMMAND == "bash" ]] || [[ $PARENT_SCRIPT_COMMAND == "-bash" ]] || [[ $PARENT_SCRIPT_COMMAND == "" ]]
+	then
+        RUNNING_SUBPROCESSES=$(echo "$SUBPROCESSES_PID_TEXT" | grep -v "$SCRIPT_COMMAND" | awk '{print $1}')
+    else
+        RUNNING_SUBPROCESSES=$(echo "$SUBPROCESSES_PID_TEXT" | grep -v "$SCRIPT_COMMAND" | grep -v "$PARENT_SCRIPT_COMMAND" | awk '{print $1}')
+    fi
+}
+
 function kill_subprocesses() 
 {
     # kills only subprocesses of the current process
     #pkill -15 -P $$
     #kill -15 $(pgrep -P $$)
+    #echo "killing processes..."
     
     # kills all descendant processes incl. process-children and process-grandchildren
-    # option 1
-    RUNNING_SUBPROCESSES=$(pgrep -g $(ps -o pgid= $$))
-    kill -15 $RUNNING_SUBPROCESSES
-    wait $RUNNING_SUBPROCESSES 2>/dev/null
-    # option 2
-    #{ kill -15 $RUNNING_SUBPROCESSES && wait $RUNNING_SUBPROCESSES; } >/dev/null 2>&1
-    # option 3
-    #kill -13 $RUNNING_SUBPROCESSES
+    # giving subprocesses the chance to terminate cleanly kill -15
+    get_running_subprocesses
+    if [[ $RUNNING_SUBPROCESSES != "" ]]
+    then
+        kill -15 $RUNNING_SUBPROCESSES
+        # do not wait here if a process can not terminate cleanly
+        #wait $RUNNING_SUBPROCESSES 2>/dev/null
+    else
+        :
+    fi
+    # waiting for clean subprocess termination
+    TIME_OUT=0
+    while [[ $RUNNING_SUBPROCESSES != "" ]] && [[ $TIME_OUT -lt 3 ]]
+    do
+        get_running_subprocesses
+        sleep 1
+        TIME_OUT=$((TIME_OUT+1))
+    done
+    # killing the rest of the processes kill -9
+    get_running_subprocesses
+    if [[ $RUNNING_SUBPROCESSES != "" ]]
+    then
+        kill -9 $RUNNING_SUBPROCESSES
+        wait $RUNNING_SUBPROCESSES 2>/dev/null
+    else
+        :
+    fi
+    # unsetting variable
     unset RUNNING_SUBPROCESSES
 }
 
@@ -792,10 +826,22 @@ function stop_sudo() {
     sudo -k
 }
 
-#trap "unset SUDOPASSWORD; printf '\n'; echo 'killing subprocesses...'; kill_subprocesses >/dev/null 2>&1; echo 'done'; echo 'killing main process...'; kill_main_process" SIGHUP SIGINT SIGTERM
-trap "unset_variables; printf '\n'; kill_subprocesses >/dev/null 2>&1; kill_main_process" SIGHUP SIGINT SIGTERM
-# kill main process only if it hangs on regular exit
-trap "unset_variables; kill_subprocesses >/dev/null 2>&1; exit" EXIT
+### trapping
+[[ "${BASH_SOURCE[0]}" != "${0}" ]] && SCRIPT_SOURCED="yes" || SCRIPT_SOURCED="no"
+[[ $(echo $(ps -o stat= -p $PPID)) == "S+" ]] && SCRIPT_SESSION_MASTER="no" || SCRIPT_SESSION_MASTER="yes"
+# a sourced script does not exit, it ends with return, so checking for session master is sufficent
+# subprocesses will not be killed on return, only on exit
+#if [[ "$SCRIPT_SESSION_MASTER" == "yes" ]] && [[ "$SCRIPT_SOURCED" == "no" ]]
+if [[ "$SCRIPT_SESSION_MASTER" == "yes" ]]
+then
+    # script is session master and not run from another script (S on mac Ss on linux)
+    trap "printf '\n'; kill_subprocesses >/dev/null 2>&1; unset SUDOPASSWORD; kill_main_process" SIGHUP SIGINT SIGTERM
+    trap "kill_subprocesses >/dev/null 2>&1; unset SUDOPASSWORD; exit" EXIT
+else
+    # script is not session master and run from another script (S+ on mac and linux)
+    trap "printf '\n'; unset SUDOPASSWORD; kill_main_process" SIGHUP SIGINT SIGTERM
+    trap "unset SUDOPASSWORD; exit" EXIT
+fi
 #set -e
 
 # creating directory and adjusting permissions
@@ -817,20 +863,31 @@ then
 else
     echo "homebrew is installed..."
 fi
+
 # checking if homebrew-cask is installed
-if [[ $(brew cask --version | grep "caskroom/homebrew-cask") == "" ]]
+#if [[ $(brew cask --version | grep "homebrew-cask") == "" ]]
+#then
+#    echo "homebrew-cask not installed, exiting script..."
+#    exit
+#else
+#    echo "homebrew-cask is installed..."
+#fi
+#echo ''
+
+brew cask --version 2>&1 >/dev/null
+if [[ $? -eq 0 ]]
 then
+    echo "homebrew-cask is installed..."
+else
     echo "homebrew-cask not installed, exiting script..."
     exit
-else
-    echo "homebrew-cask is installed..."
 fi
 echo ''
 
 # checking if online
 echo "checking internet connection..."
 ping -c 3 google.com > /dev/null 2>&1
-if [ $? -eq 0 ]
+if [[ $? -eq 0 ]]
 then
     echo "we are online, running script..."
     echo ''
